requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.3
# Optional if you handle image uploads:
Pillow>=10.0
# Optional for API docs:
drf-spectacular>=0.27
